Textlabel: Add documentation + fix return types
